backend/requirements.txt
flask
flask-sqlalchemy
flask-bcrypt
python-dotenv
flask-session
flask-cors
flask_restful
flask_jwt_extended
flask_Restful
flask-mail
pyPDF2
nltk
spacy
scikit-learn
pandas
numpy
google-generativeai